6.8 all componet out of APP
